ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
python-pptx==0.6.23
Pillow==10.1.0
openai==1.3.7
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
nltk==3.8.1
aiohttp==3.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
